[libconnman-qt] Ignore WiFi networks with zeroed MACs. JB#56225 OMP#JOLLA-552
Set to ignore new networks with a zeroed (00:00:00:00:00:00) BSSID.
These can be saved by ConnMan in some cases because of timing with scan
results from weak networks etc. and these should not appear to UI. WiFi
networks can appear without name as well and they are saved networks
with an empty BSSID. This has a slight reduction to the processing
required for the scan results to be shown in connection selector.

Also fix one copy-paste error with ethernet services.